Improved the dot coloring
By default, there is an array of symmetries and an array of colors.
Each symmetry will be assigned a unique color. But when the user filters
by one or several symmetries, it will recolor the points to make the
distinction more clear. A legend where the meaning of the colors is defined
would be useful now.